Fix "dueDate" bug
The RTAC "dueDate" field was being set regardless of whether or not the field was set in the loan. If "dueDate" was not set in the loan, mod-rtac set it to the current date, due to the field being "null" when extracted from the loan. Now, if "dueDate" is not set in the loan, mod-rtac will not set "dueDate" in the response JSON.