MODRTAC-17: Prevent parentheses with just whitespace in volume
If an enumeration field is removed in the UI, instead of removing the
value from the json, it sets the field to an empty string. Likewise, the
enumeration fields allow white space only to be values, which will be
stored in the json. The existing code treats these as usable values and
will return the volume info as required, but as "()" or
"(<whitespaces>)".

Now, we will not return volume information for whitespace only/empty
string values, nor the parentheses that would surround the volume data.