Upgrade to RMB 34.0.0. MODRTAC-93
The upgrade doesn't work in a straight-forward manner in this case.

A bug in the logfile is exposed where the wrong file name is used.
Additional dependencies are suddenly no longer available, requiring javax.ws.rs:javax.ws.rs-api to provide them.

The aspectj is upgraded from 1.9.6 to 1.9.7 to match what RMB uses to avoid additional problems regarding this.
For whatever reason the 1.9.7 from RMB does not get automatically used and the version must be explicitly specified.

There are currently test failures that appear to be related to the RMB 34.0.0 upgrade in regards to vertx and threading.
There may be a bug in RMB or vertx being exposed here.