MODRTAC-96: Add personal data disclosure form.
This module involves Items, Loans, and Holdings but none of the personal data is being pulled in or transferred in regards to these.

The notes fields from the Holdings are being pulled in and are being mapped into the "Status" field.
These notes do get transmitted and so I consider this falls within the "stores fields with free-form text" under the prescribed definition of "store" (see mapHoldingStatements() and fromHoldingToRtacHolding() in FolioToRtacMapper.java).